% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-pru.R
\name{theme_pru}
\alias{theme_pru}
\title{Add PRU theme to ggplot chart}
\usage{
theme_pru()
}
\description{
This function allows you to add the pru theme to your ggplot graphics.
}
\examples{
library(ggplot2)

p1 <- ggplot(mpg, aes(displ, hwy, colour = class)) +
 geom_point() +
 labs(title = 'Scatterplot example in ggplot2',
      subtitle = 'A plot that is only useful for demonstration purposes')
p1 + theme_pru()

p2 <- ggplot(mpg[1:137,], aes(class)) +
 geom_bar(aes(fill = manufacturer)) +
 labs(title = 'Barplot example in ggplot2',
      subtitle = 'A plot that is only useful for demonstration purposes')
p2 + theme_pru()
}
\keyword{theme_pru}
