% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-logo.R
\name{add_logo}
\alias{add_logo}
\title{Add PRU logo and save ggplot chart}
\usage{
add_logo(plot_name, source_name,
  save_filepath = file.path(Sys.getenv("TMPDIR"), "tmp-nc.png"),
  width_pixels = 640, height_pixels = 450,
  logo_image_path = "data/ectl-logo.png")
}
\arguments{
\item{plot_name}{The variable name of the plot you have created that you want to format and save}

\item{source_name}{The text you want to come after the text 'Source:' in the bottom left hand side of your side}

\item{save_filepath}{Exact filepath that you want the plot to be saved to}

\item{width_pixels}{Width in pixels that you want to save your chart to - defaults to 640}

\item{height_pixels}{Height in pixels that you want to save your chart to - defaults to 450}

\item{logo_image_path}{File path for the logo image you want to use in the right hand side of your chart,
which needs to be a PNG file - defaults to PRU logo that sits within the data folder of your package}
}
\value{
(Invisibly) an updated ggplot object.
}
\description{
Running this function will add the PRU logo in your plot and save it to your specified location.
}
\examples{
library(ggplot2)

myplot <- ggplot(mpg, aes(displ, hwy, colour = class)) +
geom_point() +
theme_pru()

add_logo(plot_name = myplot,
source = "The source for my data",
save_filepath = "filename_that_my_plot_should_be_saved_to-nc.png",
width_pixels = 640,
height_pixels = 450,
logo_image_path = "data/ectl-logo.png"
)

}
\keyword{add_logo}
