% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-pru.R
\name{theme_pru}
\alias{theme_pru}
\title{Add PRU theme to ggplot chart}
\usage{
theme_pru()
}
\description{
This function allows you to add the pru theme to your ggplotgraphics.
}
\examples{
ggplot(mpg, aes(displ, hwy, colour = class)) +
geom_point() +
theme_pru()
}
\keyword{theme_pru}
