% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-pru.R
\name{theme_pru}
\alias{theme_pru}
\title{Add PRU theme to ggplot chart}
\usage{
theme_pru()
}
\description{
This function allows you to add the pru theme to your ggplot graphics.
}
\examples{
library(ggplot2)

p1 <- ggplot(mpg, aes(displ, hwy, colour = class)) +
 geom_point() +
 theme_pru()
p1

p1 + theme(panel.grid.major.y = ggplot2::element_blank(),
           axis.line.y = ggplot2::element_line(colour = "#D5D5D5",
                                               size = 1.2,
                                               linetype = "solid",
                                               lineend = "round")) +
 labs(title = 'Scatterplot example in ggplot2',
      subtitle = 'A plot that is only useful for demonstration purposes')

p2 <- ggplot(mpg[1:137,], aes(class)) +
 geom_bar(aes(fill = manufacturer)) +
 theme_pru() +
 labs(title = 'Barplot example in ggplot2',
      subtitle = 'A plot that is only useful for demonstration purposes')
p2
}
\keyword{theme_pru}
